# public/openapi.yaml
openapi: 3.1.0
info:
  title: Cimbra API v1
  version: '2026-09-01'
  description: Versioned financial infrastructure API backed by a persistent double-entry ledger, scoped S2S API keys and signed durable webhooks. This deployment is a sandbox and does not move real funds. Every response includes X-Request-Id and Cimbra-Version; responses after successful API-key authorization also expose X-RateLimit-Limit, X-RateLimit-Remaining and X-RateLimit-Reset. Mutations replayed from an idempotency key include Idempotent-Replayed, while error responses may include Cimbra-Should-Retry so official SDKs can make safe retry decisions.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
...
servers:
  - url: https://cimbra-rose.vercel.app
    description: Persistent sandbox. Does not move real funds.
security: []
paths:
...
  /api/v1/instant-transfers:
    get:
      operationId: listInstantTransfers
      summary: List sandbox instant transfers
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
        - name: scheme
          in: query
          schema: { type: string, enum: [credit_push, debit_pull, qr_collect] }
      responses:
        '200': { description: Instant transfer page }
        '403': { description: Missing transfers:read scope or console role }
    post:
      operationId: createInstantTransfer
      summary: Create a sandbox credit push with holder confirmation
      description: Internal CVU/alias settles between product accounts. External CBU/CVU posts cash-out to settlement. Inbound credits a Cimbra CVU from an external originator. Rail is cimbra_sandbox, not Coelsa.
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [externalReference, accountId, destination, description, amount, currency, confirmHolder, holderName, taxIdLast4]
              properties:
                externalReference: { type: string, minLength: 2, maxLength: 100 }
                accountId: { type: string, format: uuid }
                destination: { type: string, minLength: 6, maxLength: 22 }
                description: { type: string, minLength: 2, maxLength: 180 }
                amount:
                  oneOf:
                    - { type: string, pattern: '^\d+(?:\.\d+)?$' }
                    - { type: number, exclusiveMinimum: 0, maximum: 10000000 }
                currency: { type: string, enum: [ARS] }
                direction: { type: string, enum: [outbound, inbound] }
                confirmHolder: { type: boolean }
                holderName: { type: string, minLength: 2, maxLength: 160 }
                taxIdLast4: { type: string, pattern: '^\d{4}$' }
                signals: { $ref: '#/components/schemas/RiskSignalsInput' }
      responses:
        '201': { description: Instant transfer created }
        '200': { description: Idempotent replay }
        '400': { description: Invalid payload }
        '403': { description: Missing transfers:write scope or finance.write permission }
        '404': { description: Account, alias or sandbox CVU not found }
        '409': { description: Inactive account, currency mismatch or idempotency conflict }
        '422': { description: Holder mismatch, insufficient funds or risk decline }
...
  /api/v1/debit-requests:
    get:
      operationId: listDebitRequests
      summary: List internal debit-pull requests
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200': { description: Debit request page }
        '403': { description: Missing transfers:read scope or console role }
    post:
      operationId: createDebitRequest
      summary: Request an internal debit from another Cimbra account in the tenant
      description: External CBU/CVU debit is rejected with external_debit_not_supported. This is not DEBIN over Coelsa.
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [externalReference, collectorAccountId, payerDestination, description, amount, currency]
              properties:
                externalReference: { type: string, minLength: 2, maxLength: 100 }
                collectorAccountId: { type: string, format: uuid }
                payerDestination: { type: string, minLength: 6, maxLength: 22 }
                description: { type: string, minLength: 2, maxLength: 180 }
                amount:
                  oneOf:
                    - { type: string, pattern: '^\d+(?:\.\d+)?$' }
                    - { type: number, exclusiveMinimum: 0, maximum: 10000000 }
                currency: { type: string, enum: [ARS] }
                expiresInMinutes: { type: integer, minimum: 5, maximum: 1440 }
      responses:
        '201': { description: Debit request created pending acceptance }
        '200': { description: Idempotent replay }
        '400': { description: Invalid payload }
        '403': { description: Missing transfers:write scope or finance.write permission }
        '422': { description: External debit is not supported }
  /api/v1/debit-requests/{id}/respond:
    post:
      operationId: respondDebitRequest
      summary: Accept or reject a pending internal debit request
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [decision]
              properties:
                decision: { type: string, enum: [accept, reject] }
                signals: { $ref: '#/components/schemas/RiskSignalsInput' }
      responses:
        '201': { description: Debit accepted and settled, or rejected }
        '200': { description: Idempotent replay }
        '403': { description: Missing transfers:write scope or finance.write permission }
        '404': { description: Debit request not found }
        '409': { description: Request is not pending or expired }
        '422': { description: Insufficient funds or risk decline }
...
  /api/v1/payment-links/{id}/pay:
    post:
      operationId: payPaymentLink
      summary: Collect a sandbox payment link
      description: internal debits another Cimbra account in the tenant. sandbox_inbound posts a cash-in against settlement. Card acquiring is not supported.
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [method]
              properties:
                method: { type: string, enum: [internal, sandbox_inbound, card, pos, tap_to_phone, qr_interoperable] }
                payerAccountId: { type: string, format: uuid }
                signals: { $ref: '#/components/schemas/RiskSignalsInput' }
      responses:
        '201': { description: Payment link collected }
        '200': { description: Idempotent replay }
        '400': { description: Invalid payload }
        '403': { description: Missing payments:write scope or finance.write permission }
        '404': { description: Payment link not found }
        '409': { description: Link not open or already paid }
        '422': { description: Unsupported method, insufficient funds or risk decline }
...
components:
  parameters:
    ResourceId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
    PageLimit:
      in: query
      name: limit
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    PageCursor:
      in: query
      name: cursor
      required: false
      description: Opaque cursor returned as nextCursor by the previous page.
      schema: { type: string }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: Stable unique key for this logical write. Replays return the original resource; reusing it with a different payload returns 409.
      schema: { type: string, minLength: 8, maxLength: 100, pattern: '^[A-Za-z0-9._:-]+$' }
  schemas:
...
    RiskSignalsInput:
      type: object
      additionalProperties: false
      properties:
        deviceReference: { type: string, minLength: 2, maxLength: 160, writeOnly: true }
        identityReference: { type: string, minLength: 2, maxLength: 160, writeOnly: true }
        deviceTrust: { type: string, enum: [trusted, unknown, suspicious] }
        identityVerified: { type: boolean }
        ipCountry: { type: string, pattern: '^[A-Za-z]{2}$' }
        countryMismatch: { type: boolean }
...
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-cimbra_session
      description: First-party opaque session token. Only its SHA-256 hash is stored server-side.
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: Cimbra API key
      description: S2S key with a cim_sk_test_ prefix. The plaintext secret is shown once and only its SHA-256 hash is stored.